feat: Add ThemeProvider, Header, Nav, Outlet and GlobalStyle components to App
This commit adds the ThemeProvider component, which is used to provide a theme object to all styled components in the app. It also adds the Header, Nav, and Outlet components, which make up the main structure of the app. Lastly, the GlobalStyle component is added to apply global styles to the app.